# Conflicts:
#	game_artifact/src/main/java/tr/edu/ku/comp302/App.java
#	game_artifact/src/main/java/tr/edu/ku/comp302/domain/entity/FireBall.java
#	game_artifact/src/main/java/tr/edu/ku/comp302/domain/entity/Lance.java
#	game_artifact/src/main/java/tr/edu/ku/comp302/domain/handler/BarrierRenderer.java
#	game_artifact/src/main/java/tr/edu/ku/comp302/domain/handler/BuildHandler.java
#	game_artifact/src/main/java/tr/edu/ku/comp302/domain/handler/ImageHandler.java
#	game_artifact/src/main/java/tr/edu/ku/comp302/domain/handler/LevelHandler.java
#	game_artifact/src/main/java/tr/edu/ku/comp302/domain/handler/collision/CollisionHandler.java
#	game_artifact/src/main/java/tr/edu/ku/comp302/domain/lanceofdestiny/LanceOfDestiny.java
#	game_artifact/src/main/java/tr/edu/ku/comp302/domain/lanceofdestiny/Level.java
#	game_artifact/src/main/java/tr/edu/ku/comp302/ui/frame/MainFrame.java
#	game_artifact/src/main/java/tr/edu/ku/comp302/ui/panel/BuildPanelModel.java
#	game_artifact/src/main/java/tr/edu/ku/comp302/ui/panel/LevelPanel.java
#	game_artifact/src/main/java/tr/edu/ku/comp302/ui/view/View.java

+++ game_artifact/src/main/java/tr/edu/ku/comp302/domain/handler/DatabaseHandler.java
@@ -343,7 +343,7 @@ public class DatabaseHandler {
         String query = "SELECT * FROM Save WHERE id = ?";
         FireballData fireball = null;
         LanceData lance = null;
-        double score = 0.0;
+        int score = 0;
         try (Connection connection = getConnection()) {
             assert connection != null;
             try (PreparedStatement ps = connection.prepareStatement(query)) {
@@ -352,7 +352,7 @@ public class DatabaseHandler {
                     if (rs.next()) {
                         fireball = new FireballData(rs.getDouble("fireball_x"), rs.getDouble("fireball_y"), rs.getDouble("fireball_dx"), rs.getDouble("fireball_dy"));
                         lance = new LanceData(rs.getDouble("lance_x"), rs.getDouble("lance_y"), rs.getDouble("lance_angle"));
-                        score = rs.getDouble("score");
+                        score = rs.getInt("score");
                     }
                 }
             }
